pyproject.toml
# pyproject.toml
# =========================================
# Project: nutri-matic 🍹
# =========================================
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nutri-matic"
version = "0.1.1"
description = "Cookiecutter utilities for jcook3701's cookiecutter templates."
authors = [
  { name = "Jared Cook", email = "jcook3701+github@gmail.com" }
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dependencies = [
  "cookiecutter>=2.6",
  "Deprecated>=1.3",
  "GitPython>=3.1",
  "pydantic>=2.12",
  "sphinx>=8.0",
  "toml>=0.10",
  "typer>=0.20",
  "types-Deprecated>=1.3",
  "types-requests>=2.32",
  "types-PyYAML>=6.0",
  "types-toml>=0.10",
] 
keywords = [
  "python3",
  "cookiecutter",
  "templates",
  "utilities",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
[project.urls]
Homepage = "https://github.com/jcook3701/nutri-matic"
Repository = "https://github.com/jcook3701/nutri-matic"
Documentation = "https://github.com/jcook3701/nutri-matic/docs/jekyll"

[project.optional-dependencies]
dev = [
  "build>=1.3",
  "bump-my-version>=1.2",
  "coverage>=7.12",
  "black>=25.11",
  "ruff>=0.14.0",
  "mypy==1.9",
  "myst-parser>=0.18",
  "nameisok>=0.1",
  "pytest>=8.0",
  "pytest-cov>=7.0",
  "yamllint>=1.0",
  "twine>=6.2",
]
docs = [
  "sphinx>=8.0",
  "sphinx-rtd-theme>=3.0",
  "sphinx-markdown-builder>=0.6",
  "sphinx-autodoc-typehints>=3.5",
]
# ====================================
# Optional entry point:
# lets you run with either:
#   1. ```python -m nutrimatic```
#   2. ```nutrimatic```
# ====================================
[project.scripts]
nutrimatic = "nutrimatic.cli.main:app"
nm-util = "nutrimatic.cli.main:app"
nm-build = "nutrimatic.cli.build:app"
nm-cfg = "nutrimatic.cli.config:app"
nm-templates = "nutrimatic.cli.templates:app"

# ====================================
# Setuptools Configuration:
# ====================================
[tool.setuptools.packages.find]
where = [
  "src",
  ]
include = [
  "nutrimatic*",
]
# ====================================
# Version Bumping (bump-my-version):
# ====================================
[tool.bumpversion]
current_version = "0.1.1"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
tag = false
commit = false

[[tool.bumpversion.files]]
filename = "Makefile"
[[tool.bumpversion.files]]
filename = "./docs/jekyll/Makefile"
[[tool.bumpversion.files]]
filename = "./docs/jekyll/_config.yml"
[[tool.bumpversion.files]]
filename = "./docs/sphinx/conf.py"
# ====================================
# Ruff Configuration: (linter, formatter, import sorter)
# ====================================
[tool.ruff]
target-version = "py311"
# Directories to lint and format
src = [
  "src",
  "tests",
]
# Exclude these paths
exclude = [
  ".venv",
  "build",
  "dist",
  "__pycache__",
  "docs",
  "*.egg-info",
  ".git",
  ".github",
]
# Maximum line length
line-length = 120
# ====================================
# Ruff Lint Configuration:
# ====================================
[tool.ruff.lint]
fixable = ["ALL"]
# Enable linting rules
select = [
  "W",  # PyCodeStyle warnings
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort (import sorting)
  "N",  # pep8-naming
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "A",  # flake8-builtins
  "C4", # comprehensions
  "TID", # tidy imports
  "T20", # print statements
  "FA", # future annotations
  "PL", # pylint-style checks
  "RUF", # ruff-specific
]
# Ignore specific rules
ignore = [
  "E501",  # line too long (handled by formatter)
  "PLR2004",  # magic value comparison
  "B008",  # function call in default argument
]
# ====================================
# Import Sorting Configuration: (Python helper scripts)
# ====================================
[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["nutrimatic"]
# ====================================
# Formatter Configuration:
# Ruff also formats YAML, TOML, and Markdown
# ====================================
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
docstring-code-format = true
# ====================================
# Black Configuration: (Formatting)
# ====================================
[tool.black]
line-length = 88          # Maximum line length (default is 88)
target-version = ["py310"] # Target Python version(s)
skip-string-normalization = false  # Whether to skip normalizing string quotes
include = '\.pyi?$'       # Which files to include (regex)
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
# ====================================
# Mypy (type-checker) Configuration:
# ====================================
[tool.mypy]
python_version = 3.11
mypy_path = "src"
strict = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
# ====================================
# Pytest Configuration:
# ====================================
[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-v", "--maxfail=1", "-W always", "-rw", "--strict-markers"]
testpaths = ["tests"]
pythonpath = ["src"]